+++ src/core/impl/array/aligned_dynamic_array.h
@@ -7,7 +7,7 @@ namespace ark {
 
 template<typename T, typename ALIGNMENT = T> class AlignedDynamicArray : public Array<T> {
 public:
-    AlignedDynamicArray(uint32_t length)
+    AlignedDynamicArray(size_t length)
         : _data(new ALIGNMENT[sizeof(T) * length / sizeof(ALIGNMENT)]), _ptr(reinterpret_cast<T*>(_data)), _length(length) {
         DCHECK((sizeof(T) * length) % sizeof(ALIGNMENT) == 0, "length = %d, alignment = %d", length, sizeof(ALIGNMENT));
     }
@@ -16,7 +16,7 @@ public:
         delete[] _data;
     }
 
-    virtual uint32_t length() override {
+    virtual size_t length() override {
         return _length;
     }
 
@@ -27,7 +27,7 @@ public:
 private:
     ALIGNMENT* _data;
     T* _ptr;
-    uint32_t _length;
+    size_t _length;
 };
 
 }
